requirements.txt
Django
django-filter
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg